13 March 2024

+++ src/main/java/com/chunkie/live_lyrics_server/entity/Song.java
@@ -14,6 +14,6 @@ public class Song {
 
     private String songAlbum;
 
-    private Long songDuration;
+    private Integer songDuration;
 
 }
